\name{Annotate.ff}
\alias{Annotate.ff}
\title{Annotation of segmented regions and generation of an HTML QC report, from a RDS file.}
\usage{
  Annotate.ff(RDS.file = NULL, ...)
}
\arguments{
  \item{RDS.file}{File generated by \code{Segment.ASCAT()}.}
  \item{...}{Any parameter for the function \code{Annotate()}, except \code{data}.}
}
\description{
  This function performs the genomic annotation of segmented regions, and builds an HTML interactive QC report, from a RDS file.
}
\author{
  Bastien Job
}

