% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_builder.R
\name{annotateRDS}
\alias{annotateRDS}
\title{Annotate CNV}
\usage{
annotateRDS(fit.val, sample, segmenter, build = "hg19",
  bin.size = 50000, feature.set = c("bins"), ...)
}
\arguments{
\item{fit.val}{A dataframe of the *.gammaEval.txt file}

\item{sample}{Sample name}

\item{segmenter}{Segmenter used (only works with ASCAT)}

\item{build}{Genome build (only works with hg19) [Default=hg19]}

\item{bin.size}{Bin-size for feature creation [Default=50000]}

\item{...}{}
}
\value{
Returns a list containing 3 elements: 'seg', 'genes', and 'bins'
for annotated datastructure fo a seg file, genes per row, and bins per row
}
\description{
This function annotates the output of the ASCN ASCAT file. It scans
the output directory for the *.gammaEval.txt to select the best "Goodness of Fit"
score, then load in the corresponding RDS file.  Using this RDS file, it extracts
and calculates the absolute allele-specific copy-number (modalA, modalB, modalAB),
as well as the L2R copy-ratios associated with them (nAraw, nBraw, seg.mean). It 
then annotates each segment and attributes a value to each gene in the knownGenes 
data structure
}
\examples{
 annotateRDS(fit.val.df, 'SampleX', 'ASCAT', build='hg19', bin.size=50000)
}
