% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDS_builder.R
\name{Build.OMNI25}
\alias{Build.OMNI25}
\title{Build.OMNI25}
\usage{
Build.OMNI25(illumina.dir, parent.dir)
}
\arguments{
\item{illumina.dir}{}

\item{parent.dir}{}
}
\value{

}
\description{
Build.OMNI25
}
\examples{
Build.OMNI25(illumina.dir='/mnt/work1/users/pughlab/projects/cancer_cell_lines/GNE/illumina/GNE_Matrices',
eacon.dir='/mnt/work1/users/pughlab/projects/cancer_cell_lines/GNE')
}
