% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDS_builder.R
\name{Build.SNP6}
\alias{Build.SNP6}
\title{Build.SNP6}
\usage{
Build.SNP6(ao.df, meta.b, affy.meta, CEL, sex.chr, samplename, cs,
  num.cores = 4)
}
\arguments{
\item{ao.df}{}

\item{meta.b}{}

\item{affy.meta}{}

\item{CEL}{}

\item{sex.chr}{}

\item{samplename}{}

\item{cs}{}
}
\description{
Build.SNP6
}
