% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_builder.R
\name{buildCbioOut}
\alias{buildCbioOut}
\title{cBio-formatted table builder}
\usage{
buildCbioOut(
  gr.cnv,
  cbio.path = "./out/cBio",
  pattern = "_CNA",
  cbio.cna.file = NULL,
  cbio.linear.file = NULL,
  cbio.seg.file = NULL,
  cbio.RAWseg.file = NULL,
  amp.thresh = 5,
  add.on.to.existing = TRUE,
  ...
)
}
\arguments{
\item{gr.cnv}{List output from annotateRDS.Batch() function}

\item{cbio.path}{Relative path to the cBio output path}

\item{pattern}{Regex pattern for existing linear_CNA and _CNA files [Default="_CNA"]}

\item{cbio.cna.file}{Names of existing data_CNA.txt files [Defaul=NULL]}

\item{cbio.linear.file}{Names of existing data_linear_CNA.txt files [Defaul=NULL]}

\item{cbio.seg.file}{Names of existing data_cna_hg19.seg files [Defaul=NULL]}

\item{amp.thresh}{Absolute CN to start calling AMP from GAINs [Default=5]}

\item{add.on.to.existing}{Builds on existing cBio data found [Default=TRUE]}

\item{...}{}

\item{overwrite}{Sample ID to overwrite in cBio output if already existing [Default=NULL]}
}
\description{
Generates cBioportal-style formatted tsv's and seg files as per
details from https://docs.cbioportal.org/5.1-data-loading/data-loading/file-formats
}
\examples{
    buildCbioOut(gr.cnv, cbio.path="./out/cBio", overwrite=c('YT_4941', '5637_3858'))
}
