% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_builder.R
\name{buildPSetOut}
\alias{buildPSetOut}
\title{PSet builder for PharmacoGX}
\usage{
buildPSetOut(gr.cnv, anno.name, pset.path, cols = c("seg.mean", "nAraw",
  "nBraw", "nMinor", "nMajor", "TCN"), ...)
}
\arguments{
\item{gr.cnv}{List output from annotateRDS.Batch() function}

\item{anno.name}{The ID of the group you are analyzing (e.g. 'GDSC')}

\item{pset.path}{Relative path to the PSet output path}

\item{cols}{The columns from gr.cnv to build into assayData() objects}

\item{...}{}

\item{meta}{A dataframe where one column contains the sample IDs found in gr.cnv}
}
\description{
Builds the PSet that is used in PharmacoGX.  It is defined to interface
with the output from annotateRDS.Batch() function
}
\examples{
    buildPSetOut(gr.cnv, "CGP", pset.path, meta=cell.line.anno)
}
