% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccle.meta}
\alias{ccle.meta}
\title{CCLE Metadata}
\format{
A dataframe containing CCLE metadata:
}
\source{
\href{https://data.broadinstitute.org/ccle/Cell_lines_annotations_20181226.txt}{CCLE metadata}
CCLE_sample_info_file_2012-10-18.txt
Cell_lines_annotations_20181226.txt
}
\usage{
data(CCLE_meta)
}
\description{
CCLE Metadata
}
\examples{
data(CCLE_meta)
}
\keyword{datasets}
