\name{EaCoN.Annotate.ff}
\alias{EaCoN.Annotate.ff}
\title{Annotation of segmented regions and generation of an HTML QC report, from a RDS file.}
\usage{
  EaCoN.Annotate.ff(RDS.file = NULL, ...)
}
\arguments{
  \item{RDS.file}{File generated by \code{EaCoN.Segment()}.}
  \item{...}{Any parameter for the function \code{EaCoN.Annotate()}, except \code{data}.}
}
\description{
  This function performs the normalization of the L2R and BAF signals from a data object obtained with \code{EaCoN.WES.Bin()} (or \code{EaCoN.WES.Bin.Batch()}).
}
\author{
  Bastien Job
}

