\name{EaCoN.WES.Normalize.ff.Batch}
\alias{EaCoN.WES.Normalize.ff.Batch}
\title{Normalization of binned WES data from a RDS file in batch mode, with multithreading.}
\usage{
  EaCoN.WES.Normalize.ff.Batch(BIN.RDS.files = list.files(path = getwd(),
  pattern = "_binned.RDS$", all.files = FALSE, full.names = TRUE, recursive = TRUE,
  ignore.case = FALSE, include.dirs = FALSE), nthread = 1, cluster.type = "PSOCK", ...)
}
\arguments{
  \item{BIN.RDS.files}{A list of RDS files generated by \code{EaCoN.WES.Bin()}.}
  \item{nthread}{Number of threads.}
  \item{cluster.type}{Type of multithreading cluster (see \code{?parallel::makeCluster()}).}
  \item{...}{Any parameter for the function \code{EaCoN.WES.Normalize()}, except \code{data}.}
}
\description{
  This function is just a wrapper to \code{EaCoN.WES.Normalize.ff()} with multithreading).
}
\value{
  For each RDS file in input, a RDS file that can be used with \code{EaCoN.Segment.ff()}.
}
\author{
  Bastien Job
}

