\name{EaCoN.WES.Normalize.ff}
\alias{EaCoN.WES.Normalize.ff}
\title{Normalization of binned WES data from a RDS file.}
\usage{
EaCoN.WES.Normalize.ff(BIN.RDS.file = NULL, ...)
}
\arguments{
\item{BIN.RDS.file}{File generated by \code{EaCoN.WES.Bin()}.}
\item{...}{Any parameter for the function \code{EaCoN.WES.Normalize()}, except \code{data}.}
}
\description{
This function performs the normalization of the L2R and BAF signals from a data object obtained with \code{EaCoN.WES.Bin()} (or \code{EaCoN.WES.Bin.Batch()}).
}
\value{
A RDS file that can be used with \code{EaCoN.Segment.ff()}.
}
\author{
Bastien Job
}

