% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdsc.meta}
\alias{gdsc.meta}
\title{GDSC Metadata}
\format{A dataframe containing GDSC metadata:}
\source{
\href{https://www.cancerrxgene.org/downloads/bulk_download}{GDSC metadata}
All cell lines screened 	Cell-line-annotation
Cell_Lines_Details.xlsx
GDSC_Cell_Lines_Details.csv
}
\usage{
data(GDSC_meta)
}
\description{
GDSC Metadata
}
\examples{
data(GDSC_meta)
}
\keyword{datasets}
