\name{gen.df}
\docType{data}
\alias{gen.df}
\title{Human gene symbols and chromosomal positions (refGene)}
\description{
  This data package exists in three versions, for the hg18, hg19 and hg38 human genome builds.
}
\usage{gen.df}
\format{A dataframe with 7 columns : [1] "symbol" (gene symbol), [2] "chrom" (chromosome name in alphanumerical characters), [3] "chrN" (chromosome number in digits), [4] "start" (chromosomal location start), [5] "end" (chromosomal location end), [6] "width" (gene width), [7] "strand" (genomic strand).}
\details{This dataset is available to perform the annotation of segmentation results.}
\source{UCSC refGene}
\author{Bastien Job (bastien.job@inserm.fr)}
\keyword{datasets, object}
