\name{gen.df}
\alias{gen.df}
\alias{targetlist}
\title{Internal reference data for EaCoN}
\description{
  \code{gen.df}: A data frame containing genomic region definitions used internally by EaCoN.

  \code{targetlist}: A data frame containing a list of cancer-relevant gene targets used internally by EaCoN.
}
\keyword{internal}
