% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancan.obj.segless}
\alias{pancan.obj.segless}
\title{Pre-processed TCGA Pancancer Atlas data}
\format{
A list with three elements:
\describe{
 \item{AP-meta}{A dataframe containing metadata of all annotated samples with ploidy}
 \item{APS-meta}{A dataframe containing metadata of all annotated samples with ploidy and seg files}
 \item{breaks}{A list of 4 matrices containing binned frequencies for: 'purity', 'ploidy', 'cancerDNA fraction', 'subclonal fraction'}
}
}
\source{
\href{https://gdc.cancer.gov/about-data/publications/pancanatlas}{PanCancer Atlas}
ABSOLUTE-annotated seg file - TCGA_mastercalls.abs_segtabs.fixed.txt 
Copy Number - broad.mit.edu_PANCAN_Genome_Wide_SNP_6_whitelisted.seg
ABSOLUTE purity/ploidy file - TCGA_mastercalls.abs_tables_JSedit.fixed.txt 
Merged Sample Quality Annotations - merged_sample_quality_annotations.tsv
}
\usage{
data(pancan.obj.segless)
}
\description{
Metadata and matrices of TCGA samples that have ploidy/purity calls
on them as well, split by cancer type.
}
\details{
"The Cancer Genome Atlas (TCGA) Research Network has profiled and 
analyzed large numbers of human tumors to discover molecular aberrations 
at the DNA, RNA, protein and epigenetic levels. The resulting rich 
data provide a major opportunity to develop an integrated picture 
of commonalities, differences and emergent themes across tumor 
lineages. The Pan-Cancer Atlas initiative compares the 33 tumor 
types profiled by TCGA. Analysis of the molecular aberrations and 
their functional roles across tumor types will teach us how to extend 
therapies effective in one cancer type to others with a similar 
genomic profile."
}
\examples{
data(pancanPloidy.noSegs)
pancan.ploidy <- pancan.obj.segless$breaks$ploidy
pancan.ploidy[,c('breaks', 'ACC')]
}
\keyword{datasets}
