% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tad.gr}
\alias{tad.gr}
\title{Consensus TAD GenomicRanges}
\format{A GenomicRanges object containing condensed TADs and their jaccard index cutoff}
\usage{
data(consensusTAD)
}
\description{
Consensus TAD GenomicRanges
}
\examples{
data(consensusTAD)
}
\keyword{datasets}
