\name{targetlist}
\docType{data}
\alias{targetlist}
\title{A restricted list of human gene symbols}
\description{
  This default data package is a list of 475 gene symbols know in human cancer.
}
\usage{targetlist}
\format{A simple vector containing gene symbols.}
\details{This dataset is available to perform a subpart of the QC report.}
\source{All genes are HUGO symbols. The list itself is from a custom cancer panel at Gustave Roussy, Villejuif, FRANCE.}
\author{Bastien Job (bastien.job@inserm.fr)}
\keyword{datasets, object}
