\name{WES.Normalize.ff}
\alias{WES.Normalize.ff}
\title{Normalization of binned WES data from a RDS file.}
\usage{
WES.Normalize.ff(BIN.RDS.file = NULL, ...)
}
\arguments{
\item{BIN.RDS.file}{File generated by \code{WES.Bin()}.}
\item{...}{Any parameter for the function \code{WES.Normalize()}, except \code{data}.}
}
\description{
This function performs the normalization of the L2R and BAF signals from a data object obtained with \code{WES.Bin()} (or \code{WES.Bin.Batch()}).
}
\value{
A RDS file that can be used with \code{Segment.ASCAT.ff()}.
}
\author{
Bastien Job
}

