% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EaCoN_functions.R
\name{ASCN.ff.Batch}
\alias{ASCN.ff.Batch}
\title{ASCN.ff.Batch}
\usage{
ASCN.ff.Batch(
  RDS.files = list.files(path = getwd(), pattern = "\\\\.SEG\\\\..*\\\\.RDS$",
    full.names = TRUE, recursive = TRUE, ignore.case = TRUE, include.dirs = FALSE),
  nthread = 1,
  cluster.type = "PSOCK",
  ...
)
}
\arguments{
\item{RDS.files}{}

\item{nthread}{}

\item{cluster.type}{}

\item{...}{}
}
\value{

}
\description{
Run ASCN.ff() in batch mode
}
