% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EaCoN_functions.R
\name{Segment.ff.Batch}
\alias{Segment.ff.Batch}
\title{Run Segment.ff() in batch mode}
\usage{
Segment.ff.Batch(
  RDS.files = list.files(path = getwd(), pattern = "_processed.RDS$", full.names =
    TRUE, recursive = TRUE, ignore.case = TRUE, include.dirs = FALSE),
  segmenter = "ASCAT",
  nthread = 1,
  cluster.type = "PSOCK",
  ...
)
}
\arguments{
\item{RDS.files}{}

\item{segmenter}{}

\item{nthread}{}

\item{cluster.type}{}

\item{...}{}
}
\value{

}
\description{
Run Segment.ff() in batch mode
}
